[tp][sb][ts][dg] implemented stack growth

+++ src/userprog/exception.c
@@ -2,6 +2,7 @@
 #include <inttypes.h>
 #include <stdio.h>
 #include <vm/vm.h>
+#include <threads/vaddr.h>
 #include "userprog/gdt.h"
 #include "threads/interrupt.h"
 #include "threads/thread.h"
@@ -150,14 +151,19 @@ page_fault(struct intr_frame *f) {
     write = (f->error_code & PF_W) != 0;
     user = (f->error_code & PF_U) != 0;
 
+    void* esp = user ? f->esp : thread_current()->esp;
+
+    void *kpage = vm_handle_page_fault(fault_addr, esp);
+
     /* Kernel page faulted reading user memory. */
-    if (!user) {
+    if (!user && kpage == NULL) {
         // Set eax to 0xffffffff and copies the old value into eip.
         f->eip = * (void (**)(void)) &f->eax;
         f->eax = 0xffffffff;
         kill_process();
         NOT_REACHED();
     }
+
 
 //    /* To implement virtual memory, delete the rest of the function
 //       body, and replace it with code that brings in the page to
@@ -168,8 +174,7 @@ page_fault(struct intr_frame *f) {
 //           write ? "writing" : "reading",
 //           user ? "user" : "kernel");
 
-    void *kpage = vm_handle_page_fault(fault_addr);
-    pagedir_set_page(thread_current()->pagedir, fault_addr, kpage, true);
+
 }
 
 static void
